Agentic_AI/requirements.txt
streamlit
langchain
langchain-community
tqdm
notebook==7.1.2
openai
elasticsearch
scikit-learn
pandas
python-dotenv
matplotlib
seaborn
xlrd >= 2.0.1
unidecode
nltk
graphviz
pylance
streamlit
tabulate
notebook
nltk
langchain_groq
rouge
bert-score

## advance RAG Course

aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
annoy==1.17.3
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
build==1.2.1
cachetools==5.3.3
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.3
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.20.0
debugpy==1.8.2
decorator==5.1.1
deepdiff==7.0.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
emoji==2.12.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastembed==0.3.1
filelock==3.15.4
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
google-auth==2.30.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
kubernetes==30.1.0
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-experimental==0.0.58
langchain-openai==0.1.7
langchain-text-splitters==0.0.2
langchainhub==0.1.20
langdetect==1.0.9
langfuse==2.38.0
langsmith==0.1.82
lark==1.1.9
loguru==0.7.2
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nemoguardrails==0.9.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4 
oauthlib==3.2.2
onnx==1.16.1
onnxruntime==1.18.1
openai==1.35.7
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
ordered-set==4.1.0
orjson==3.10.5
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parso==0.8.4
pgvector==0.3.0
pillow==10.4.0
platformdirs==4.2.2
posthog==3.5.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyparsing==3.1.2
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pysbd==0.3.4
PyStemmer==2.2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
ragas==0.1.9
rapidfuzz==3.9.3
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
ruff==0.5.0
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.14.0
seaborn==0.13.2
sentence-transformers==3.0.1
shellingham==1.5.4
simpleeval==0.9.13
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.37.2
sympy==1.12.1
tabulate==0.9.0
tenacity==8.4.2
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.3
typer==0.12.3
types-requests==2.32.0.20240622
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
unstructured==0.14.9
unstructured-client==0.23.8
urllib3==2.2.2
uvicorn==0.30.1
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2


# # Windows-specific dependencies
# intel-openmp==2021.4.0; sys_platform == 'win32'
# mkl==2021.4.0; sys_platform == 'win32'
# tbb==2021.11.0; sys_platform == 'win32'

# -f https://download.pytorch.org/whl/torch_stable.html
# torch==2.3.1+cpu; sys_platform == 'win32'
# torchaudio==2.3.1+cpu; sys_platform == 'win32'
# torchvision==0.18.1+cpu; sys_platform == 'win32'


# # MacOS (Apple Silicon) specific dependencies
# -f https://download.pytorch.org/whl/torch_stable.html
# torch==2.3.1; sys_platform == 'darwin' and platform_machine == 'arm64'
# torchaudio==2.3.1; sys_platform == 'darwin' and platform_machine == 'arm64'
# torchvision==0.18.1; sys_platform == 'darwin' and platform_machine == 'arm64'

# Linux torch
-f https://download.pytorch.org/whl/torch_stable.html
torch==2.3.1+cpu; sys_platform == 'linux'
torchaudio==2.3.1+cpu; sys_platform == 'linux'
torchvision==0.18.1+cpu; sys_platform == 'linux'

## Agentic AI
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq 

